# swagger.yaml
openapi: 3.1.0
info:
  title: Job API
  description: ''
  contact: {}
  version: '1.0'
...
servers:
- url: localhost:5000/api/v1
  variables: {}
paths:
...
  /:
    get:
      tags:
      - Jobs
      summary: Get All Jobs
      description: ''
      operationId: GetAllJobs
      parameters:
      - name: User-Agent
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          examples:
          - insomnia/9.3.1
      responses:
        '200':
          description: ''
          headers: {}
          content:
            text/plain:
              schema:
                contentMediaType: text/plain
      deprecated: false
      security:
      - bearer: []
      servers:
      - url: localhost:5000/api/v1
        variables: {}
    post:
      tags:
      - Jobs
      summary: Post A Job
      description: ''
      operationId: PostAJob
      parameters:
      - name: User-Agent
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          examples:
          - insomnia/9.3.1
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/PostAJobRequest'
              - examples:
                - company: apple
                  position: intern
              contentMediaType: application/json
            example:
              company: apple
              position: intern
        required: true
      responses:
        '200':
          description: ''
          headers: {}
          content:
            text/plain:
              schema:
                contentMediaType: text/plain
      deprecated: false
      security:
      - bearer: []
      servers:
      - url: localhost:5000/api/v1
        variables: {}
    parameters: []
...
components:
  schemas:
...
    PostAJobRequest:
      title: PostAJobRequest
      required:
      - company
      - position
      type: object
      properties:
        company:
          type: string
        position:
          type: string
      examples:
      - company: apple
        position: intern
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...